requirements.txt
# Основные зависимости для Telegram бота
aiogram==3.3.0
python-dotenv==1.0.0

# Для работы с HTTP запросами (если понадобится)
aiohttp==3.9.1

# Для работы с JSON и датами
python-dateutil==2.8.2

# Для асинхронной работы с файлами
aiofiles==23.2.0 